Reduce allocations as much as possible
This change mainly changes the test types to use `&str` instead of
`String` to avoid allocations and string copying as much as possible.

This also means that we had to drop CBOR, which doesn't seem to support
zero-copy deserialization. Besides, it wasn't performing that well
anyway and not something of interest to me.